requirements.txt
django
django-environ
djangorestframework
channels
channels_redis
tensorflow>2.2
numpy